requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
google-genai==0.8.0
gunicorn==20.1.0
python-dotenv==1.0.0
google-genai